Fixed some concurrency and timing issues in downloading.
Now better handles cases when data cannot be downloaded.

Fixed two bugs occurring when using a segue to push a view controller when a link in a web view is tapped.

Much refactoring in BMAController, including extraction of a new Utilities class, NetActivityIndicatorController, which handles the spinning pinwheel icon in the upper left of the screen.

The parsing of dates now more robustly handles unparseable text.